+++ frontend/src/pages/League.tsx
@@ -1,31 +1,23 @@
-/* eslint-disable no-nested-ternary */
-/* eslint-disable camelcase */
 import 'tailwindcss/tailwind.css';
 
 import React, { ReactElement, useContext, useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom';
 
-import { FantasyPremierLeagueApi } from '../api';
 import { LeagueContext } from '../App';
 import ContentCard from '../Components/Common/ContentCard/ContentCard';
 import FormTable from '../Components/FormTable/FormTable';
 import ChartContainer from '../Components/LeagueCharts/ChartContainer';
 import LeagueInfoContainer from '../Components/LeagueInfo/LeagueInfoContainer';
 import LeagueTable from '../Components/LeagueTable/LeagueTable';
-import Loading from '../Components/Loading/Loading';
-import { IDraftPlayer, IDraftPlayerForm, IDraftPlayerStanding } from '../interfaces/League';
-import { getPlayerForm, getPlayerStandings } from '../Utils/Utils';
-
-const fantasyPremierLeagueApi = new FantasyPremierLeagueApi();
 
 function League(): ReactElement {
   const {
-    selectedSeason,
+    standings,
+    draftPlayerForms,
   } = useContext(LeagueContext);
   const [error, setError] = useState(false);
 
   useEffect(() => {
-    setError(!selectedSeason);
+    setError(!standings);
   }, []);
 
   return (
@@ -40,10 +32,10 @@ function League(): ReactElement {
         : (
           <>
             <ContentCard>
-              <LeagueTable rows={selectedSeason.standings} />
+              <LeagueTable rows={standings} />
             </ContentCard>
             <ContentCard>
-              <FormTable rows={selectedSeason.draftPlayerForms} />
+              <FormTable rows={draftPlayerForms} />
             </ContentCard>
             <ContentCard>
               <LeagueInfoContainer />